refactor EchoPlayer to take in url
