+++ frontend/src/App.jsx
@@ -11,11 +11,7 @@ import SignupForm from './components/SessionForms/SignupForm';
 import Echos from './components/Echos/Echos';
 import Profile from './components/Profile/Profile';
 import EchoCompose from './components/Echos/EchoCompose';
-
 import { getCurrentUser } from './store/session';
-
-import HowlerTest from './components/Audio/EchoPlayer';
-import WaveTest from './components/Audio/WaveSurfer';
 
 const Layout = () => {
   return (
@@ -41,14 +37,6 @@ const router = createBrowserRouter([
       {
         path: "signup",
         element: <AuthRoute component={SignupForm} />
-      },
-      {
-        path: "/audiotest",
-        element: <HowlerTest />
-      },
-      {
-        path: "/wavetest",
-        element: <WaveTest />
       },
       {
         path: "echos",
