+++ frontend/src/components/Audio/EchoPlayer.jsx
@@ -1,30 +1,44 @@
-import {Howl, Howler} from 'howler';
+import { useEffect, useRef, useState } from "react";
+import WaveSurfer from "wavesurfer.js";
+import "./EchoPlayer.css";
 
-const audioUrls = [
-    'https://teamlab-echo.s3.amazonaws.com/public/baby-shark.mp3'
-]
-  
-function HowlerTest() {
-    const babyShar = new Howl({
-        src: ['https://teamlab-echo.s3.amazonaws.com/public/baby-shark.mp3']
-    });
+function WaveTest({echoUrl}) {
+    const mySongRef = useRef(null);
+    const [songPlaying, setSongPlaying] = useState(false);
 
-    const handlePlay = (e) => {
-        e.preventDefault();
+    useEffect(() => {
+        const mySong = WaveSurfer.create({
+            container: "#myWaveForm",
+            waveColor: '#88cafb',
+            cursorWidth: 0,
+            url: echoUrl
+        })
 
-        babyShar.play();
+        mySongRef.current = mySong;
+
+        return () => {
+            mySong.destroy();
+        };
+    }, []);
+
+    const handlePlay = () => {
+        if (mySongRef.current) {
+            mySongRef.current.playPause();
+            setSongPlaying(!songPlaying);
+        }
     }
 
-    return(
+
+    return (
         <>
-        <div>
-            <button onClick={handlePlay}>Babyshark</button>
+        <div className="echoCard">
+            <div className="playContainer">
+                <button className="upperEchoButton" onClick={handlePlay}>{songPlaying ? <i className="fa-solid fa-pause"></i> : <i id="playTriangle" className="fa-solid fa-play"></i>}</button>
+                <div id="myWaveForm"></div>
+            </div>
         </div>
-        
-        
-        
         </>
-    )
+      );
 }
 
-export default HowlerTest;
+export default WaveTest;
